% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBResults.R
\name{BBResults}
\alias{BBResults}
\title{Extract Results of BetaBinomial Fitting Results}
\usage{
BBResults(Result_Dir, GenePair_IdxTable)
}
\arguments{
\item{Result_Dir}{The directory of BetaBinomial Fitting Results}

\item{GenePair_IdxTable}{The GenePair Index Table}
}
\value{
A combined Edge list of Co-perturbation Network
}
\description{
This function generates random co-perturbation Matrix. As each co-perturbation matrix is symmetric, it saves the upper_tir area of the matrix for saving the space
}
\examples{
BBResults(Result_Dir = "~/FittingResults/")
}
