% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FittingDataPreparation.R
\name{FittingDataPreparation}
\alias{FittingDataPreparation}
\title{Prepare data for Betabinomial fitting}
\usage{
FittingDataPreparation(
  FittingFile_Dir,
  Transform_CoPM_Dir,
  CoPMatrix,
  SigIdx,
  P_cutoff,
  FittingFiles
)
}
\arguments{
\item{FittingFile_Dir}{The directory for betabinomial fitting}

\item{Transform_CoPM_Dir}{The directory of transformed simulation-based Co-perturbation matrix in gene by simulation format}

\item{CoPMatrix}{The real co-perturbation matrix}

\item{SigIdx}{Index of gene pairs with significant simluation-based P-values}

\item{P_cutoff}{The P-value threshold determining significant simulation-based P-values}

\item{FittingFiles}{Number of files for following betabinomial fitting}
}
\value{
Files ready for fitting in the FittingFile_Dir
}
\description{
This function prepares fitting files for the following step of betabinomial fitting.
}
\examples{
FittingFile_Dir = "~/CopNet_Test/FittingData/", Transform_CoPM_Dir = "~/CopNet_Test/TransformedMatrix/", SigIdx = fitting_summary$qualified_gene_idx, P_cutoff = P_cutoff, FittingFiles = FittingFiles)
}
