% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenePairIdx.R
\name{GenePairIdx}
\alias{GenePairIdx}
\title{Gene pair index to gene names}
\usage{
GenePairIdx(Idx, CoMatrix)
}
\arguments{
\item{Idx}{The idx of gene pairs}

\item{CoMatrix}{The coregulation matrix}
}
\value{
The name of genes with corresponding index
}
\description{
Gene pair index to gene names
}
\examples{
GenePairIdx(Idx = fitting_summary$qualified_gene_idx, CoMatrx = CoMatrix)
}
