% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandCoRegMatrix_Generation.R
\name{RandCoRegMatrix_Generation}
\alias{RandCoRegMatrix_Generation}
\title{Generate Random Binary DEG matrix}
\usage{
RandCoRegMatrix_Generation(
  Ncores = 1,
  EachCoreSims = 1000,
  RowSumVector,
  ColSumVector,
  save_path,
  SaveSparse = FALSE
)
}
\arguments{
\item{Ncores}{Number of CPU to simulate the binary matrix with given RowSum and ColSum, and calculate the co-regulation matrix.}

\item{EachCoreSims}{Number of matrix each core simulates}

\item{RowSumVector}{The fixed RowSum Vecotor of true binary DEG matrix - Each row is a gene}

\item{ColSumVector}{The fixed ColSum Vecotor of true binary DEG matrix - Each column is an experiment}

\item{save_path}{The directory to save the simulated co-regulation matrix}

\item{SaveSparse}{Save Results in Sparse Matrix/Vector}
}
\value{
save Simulated co-regulation matrix with given rowsum and colsum of the true binary DEG matrix to the given directory
}
\description{
This function generates random co-regulation Matrix. As each co-regulation matrix is symmetric, it saves the upper_tir area of the matrix for saving the space
}
\examples{
RandCoRegMatrix_Generation(Ncores = 20,EachCoreSims = 50,RowSumVector = c(2,3,4,1,5,2,3,6,3),ColSumVector = c(4,3,2,2,2,5,3,2,3,3), save_path = "~/RandBM/")
}
