% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimP_calculation.R
\name{SimP_calculation}
\alias{SimP_calculation}
\title{Calculate simulation-based PvaluesP}
\usage{
SimP_calculation(
  transformed_CoRegM_Dir,
  save_path,
  CoTimes,
  trans_summary,
  Ncores = 1
)
}
\arguments{
\item{transformed_CoM_Dir}{The directory only contains transformed co-regulation matrix}

\item{save_path}{The directory to save the simulation-based P-value}

\item{CoTimes}{The real co-regulation times of all gene pairs}

\item{trans_summary}{result generated by function Transform_CoM_sup}

\item{Ncores}{Number of cores to run complete the task}
}
\description{
This function will calculate the simulation-based P-values based on transformed co-regulation matrix.
}
\examples{
SimP_calculation(transformed_CoRegM_Dir = "~/TransformedMatrix/", save_path = "~/SimP/", CoTimes, trans_summary, Ncores = 1)
}
