% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimP_calculation.R
\name{SimPCalculation}
\alias{SimPCalculation}
\title{Calculate simulation-based Pvalues}
\usage{
SimPCalculation(
  transformed_CoPM_Dir,
  save_path,
  CoTimes,
  trans_summary,
  Ncores = 1
)
}
\arguments{
\item{transformed_CoPM_Dir}{The directory only contains transformed co-perturbation matrix}

\item{save_path}{The directory to save the simulation-based P-value}

\item{CoTimes}{The real co-perturbation times of all gene pairs}

\item{trans_summary}{result generated by function Transform_CoPM_sup}

\item{Ncores}{Number of cores to run complete the task}
}
\value{

}
\description{
This function will calculate the simulation-based P-values based on transformed co-perturbation matrix.
}
\examples{
SimPCalculation(transformed_CoPM_Dir = "~/TransformedMatrix/", save_path = "~/SimP/", CoTimes, trans_summary, Ncores = 1)
}
