% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform_CoM.R
\name{Transform_CoM}
\alias{Transform_CoM}
\title{Transform co-regulation matrix into gene vs simulation format}
\usage{
Transform_CoM(
  RandCoM_directory,
  save_path,
  NGenePairs,
  NTransformedMatrix = 1,
  Ncores = 1
)
}
\arguments{
\item{RandCoM_directory}{The directory contains Random co-regulation matrix generated by function RandCoRegMatrix_Generation}

\item{save_path}{The directory to save the transformed co-regulation matrix}

\item{NGenePairs}{A vector contains Number of gene pairs in each transformed matrix}

\item{NTransformedMatrix}{Number of Transformed matrix splited in}

\item{Ncores}{Number of cores to complete the task}
}
\value{
Save the transformed matrix in the given directory
}
\description{
This function enables transform the gene-gene co-regulation matrix into gene by simulation format so that p values can be calculated easily when the simulation files are large
Please try Transform_CoM_sup function first with vairable NTransformedMatrix = 1 to see the size of transformed matrix
If the size is small enough (depends on the memory of your computer/server), please go with 1 transformed matrix
If the size is too large, try different NTransformedMatrix, making sure the size of each file is small.
}
\examples{
trans_summary <- Transform_CoM_sup(RandCoM_directory = "~/RandBM/", Ngenes = 10, NTransformedMatrix = 1)
Transform_CoM(RandCoM_directory = "~/RandBM/", save_path = "~/TransformedMatrix/", NGenePairs = trans_summary[,2], NTransformedMatrix = 1, Ncores = 1)
}
