% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform_CoPM.R
\name{Transform_CoPM}
\alias{Transform_CoPM}
\title{Transform co-perturbation matrix into gene vs simulation format}
\usage{
Transform_CoPM(
  RandCoPM_directory,
  save_path,
  NGenePairs,
  NTransformedMatrix = 1,
  Ncores = 1
)
}
\arguments{
\item{RandCoPM_directory}{The directory contains Random co-perturbation matrix generated by function RandCoPMatrix_Generation}

\item{save_path}{The directory to save the transformed co-perturbation matrix}

\item{NGenePairs}{A vector contains Number of gene pairs in each transformed matrix}

\item{NTransformedMatrix}{Number of Transformed matrix splited in}

\item{Ncores}{Number of cores to complete the task}
}
\value{
Save the transformed matrix in the given directory
}
\description{
This function enables transform the gene-gene co-perturbation matrix into gene by simulation format so that p values can be calculated easily when the simulation files are large
Please try Transform_CoPM_sup function first with vairable NTransformedMatrix = 1 to see the size of transformed matrix
If the size is small enough (depends on the memory of your computer/server), please go with 1 transformed matrix
If the size is too large, try different NTransformedMatrix, making sure the size of each file is small.
}
\examples{
trans_summary <- Transform_CoPM_sup(RandCoPM_directory = "~/RandBM/", Ngenes = 10, NTransformedMatrix = 1)
Transform_CoPM(RandCoPM_directory = "~/RandBM/", save_path = "~/TransformedMatrix/", NGenePairs = trans_summary[,2], NTransformedMatrix = 1, Ncores = 1)
}
