% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetaBinomialDistributionFitting.R
\name{BetaBinomialDistributionFitting}
\alias{BetaBinomialDistributionFitting}
\title{BetaBinomialDistributionFitting}
\usage{
BetaBinomialDistributionFitting(FittingFile_Dir, Result_Dir, Ncores = 1)
}
\arguments{
\item{FittingFile_Dir}{The directory contains fitting data and fitting parameters}

\item{Result_Dir}{The directory for saving the results}

\item{Ncores}{Number of cores to complete the task}
}
\value{
Result will be saved in .txt format
}
\description{
This function enables Betabinomial fitting on the simulation results.
}
\examples{
BetaBinomialDistributionFitting(FittingFile_Dir = "~/FittingData/",Result_Dir = "~/FittingResults/", Ncores = 2)
}
